Extract NodeDisabling feature trait from Dbal graph projector
